Add updates

+++ detector/scripts/clean.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
 
-# Определение пути к корневой директории проекта
 PROJECT_ROOT=$(dirname "$(dirname "$(realpath "$0")")")
-
-# Переход в корневую директорию проекта
 cd "$PROJECT_ROOT"
+echo "Перешел в директорию: $PROJECT_ROOT" | tee -a script.log
 
-# Удаление директории сборки
-rm -rf build
+if [ -d "build" ]; then
+    echo "Найдена директория build. Удаляю..." | tee -a script.log
+    rm -rf build
+    echo "Директория build удалена." | tee -a script.log
+else
+    echo "Директория build не найдена. Нечего удалять." | tee -a script.log
+fi
